add thunk for time slots
