+++ src/components/models/TimeSlotsType.model.ts
@@ -1,0 +1,4 @@
+export interface ITimeSlots {
+  doctorId: string,
+  date: string
+}
