+++ src/components/user/createAppointment/CreateAppointment.styled.ts
@@ -1,7 +1,16 @@
 import styled from 'styled-components';
+import { Form } from 'formik';
 
 export const StyledWrapper = styled.div`
   padding: 0 32px 0 0;
+`;
+
+export const StyledForm = styled(Form)`
+`;
+
+export const StyledFormContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
 `;
 
 export const StyledCrumbs = styled.div`
@@ -15,24 +24,28 @@ export const StyledSubtitle = styled.h1`
   margin: 0 0 40px 0;
 `;
 
-export const StyledFooter = styled.footer`
-  display: flex;
-  justify-content: space-between;
-`;
-
 export const StyledContainer = styled.div<{ width: string }>`
   width: 100%;
   max-width: ${(props) => (props.width ? props.width : '')};
 `;
 
-export const StyledButton = styled.button`
-  float: right;
-  border: none;
-  outline: none;
-  background: ${(props) => (!props.disabled ? '#7297FF' : '#DCE0EC')};
-  color: #ffffff;
-  border-radius: 8px;
-  line-height: 24px;
-  padding: 16px 50px;
-  cursor: pointer;
+export const StyledInputWrapper = styled.div`
+  margin-bottom: 64px;
+  position: relative;
+  &:last-of-type{
+    margin-bottom: 80px;
+  }
+`;
+
+export const StyledLabel = styled.label`
+  font-size: 13px;
+  font-weight: 500;
+  line-height: 17px;
+  position: absolute;
+  top: -25px;
+`;
+
+export const StyledBtnContainer = styled.div`
+  display: flex;
+  justify-content: flex-end;
 `;
