+++ src/components/user/createAppointment/CreateAppointment.tsx
@@ -1,42 +1,38 @@
-import React, { useEffect, useState } from 'react';
+import { Formik, Field } from 'formik';
 import { useHistory } from 'react-router-dom';
-
-import ReactCalendar from './Calendar/Calendar';
+import Calendar from './Calendar/Calendar';
 import {
-  StyledButton,
-  StyledContainer, StyledCrumbs,
-  StyledFooter,
-  StyledSubtitle,
-  StyledWrapper,
+  StyledBtnContainer,
+  StyledContainer,
+  StyledCrumbs,
+  StyledForm,
+  StyledFormContainer,
+  StyledInputWrapper, StyledLabel, StyledSubtitle, StyledWrapper,
 } from './CreateAppointment.styled';
-import Breadcrumb from './BreadCrumbs/Breadcrumb';
-import Title from './title/Title';
-import FormPatient from './Form/FormPatient';
+import SelectOccupations from './select/SelectOccupation';
+import Input from './input/Input';
+import SelectDoctor from './select/SelectDoctor';
 import Checkboxes from './checkboxes/Checkboxes';
-import useActions from '../../../hooks/useActions';
+import Breadcrumb from './BreadCrumbs/Breadcrumb';
 import routes from '../../../routes/routes';
+import { validationSchema } from './schema/validationSchema';
+import Title from './title/Title';
+import Button from './button/Button';
+import useActions from '../../../hooks/useActions';
 
-type FormValuesType = {
-  doctorId: string,
-  reason?: string,
-  note?: string
-};
+interface IFormValues {
+  date: string;
+  time: string;
+  occupation: string;
+  doctorName: string;
+  reason: string;
+  note: string;
+}
 
 const CreateAppointment = () => {
   const history = useHistory();
   const { createAppointment } = useActions();
-
   const crumbs = ['Doctors', 'Make an appointment'];
-  const [calendarData, setCalendarData] = useState<string>();
-  const [time, setTime] = useState<string>();
-  const [formData, setFormData] = useState<FormValuesType>({} as FormValuesType);
-  const [disabled, setDisabled] = useState(true);
-
-  useEffect(() => {
-    if (calendarData && time && formData!.doctorId && formData!.reason) {
-      setDisabled(false);
-    }
-  }, [calendarData, time, formData]);
 
   const selected = (crumb: string) => {
     if (crumb === 'Doctors') {
@@ -44,57 +40,75 @@ const CreateAppointment = () => {
     }
   };
 
-  const calendarHandler = (calendarDataValue: string) => {
-    setCalendarData(calendarDataValue);
+  const initialValues = {
+    date: '',
+    time: '',
+    occupation: '',
+    doctorName: '',
+    reason: '',
+    note: '',
   };
 
-  const timeHandler = (timeValue: string) => {
-    setTime(timeValue);
-  };
-
-  const formDataHandler = (formValues: FormValuesType) => {
-    if (formValues.doctorId) {
-      setFormData(formValues);
-    }
-  };
-
-  const onCreateHandler = () => {
+  const submitHandler = (values: IFormValues) => {
     const obj = {
-      date: time,
-      reason: formData!.reason,
-      note: formData!.note,
-      doctorID: formData!.doctorId,
+      date: values.time,
+      reason: values.reason,
+      note: values.note,
+      doctorID: values.doctorName,
     };
 
     createAppointment(obj);
   };
 
   return (
-    <StyledWrapper data-testid="test">
-      <StyledCrumbs>
-        <Breadcrumb crumbs={crumbs} selected={selected} />
-      </StyledCrumbs>
-      <StyledSubtitle>Make an appointment</StyledSubtitle>
-      <StyledFooter>
-        <StyledContainer width="400px">
-          <Title last={false} text="Choose a day for an appointment" number="1" />
-          <ReactCalendar onChangeData={calendarHandler} />
-        </StyledContainer>
-        <StyledContainer width="464px">
-          <Title last={false} text="Select an available timeslot" number="2" />
-          <Checkboxes
-            onChangeTime={timeHandler}
-            dataIso={calendarData}
-            formData={formData}
-          />
-        </StyledContainer>
-        <StyledContainer width="624px">
-          <Title last text="Select a doctor and define the reason of your visit" number="3" />
-          <FormPatient onChangeFormData={formDataHandler} />
-          <StyledButton disabled={disabled} onClick={onCreateHandler} data-testid="submit">Submit</StyledButton>
-        </StyledContainer>
-      </StyledFooter>
-    </StyledWrapper>
+    <Formik
+      initialValues={initialValues}
+      onSubmit={submitHandler}
+      validationSchema={validationSchema}
+    >
+      {({ setFieldValue }) => (
+        <StyledWrapper>
+          <StyledCrumbs>
+            <Breadcrumb crumbs={crumbs} selected={selected} />
+          </StyledCrumbs>
+          <StyledSubtitle>Make an appointment</StyledSubtitle>
+          <StyledForm>
+            <StyledFormContainer>
+              <StyledContainer width="400px">
+                <Title last={false} text="Choose a day for an appointment" number="1" />
+                <Field name="date" id="date" component={Calendar} />
+              </StyledContainer>
+              <StyledContainer width="464px">
+                <Title last={false} text="Select an available timeslot" number="2" />
+                <Field name="time" id="time" component={Checkboxes} />
+              </StyledContainer>
+              <StyledContainer width="624px">
+                <Title last text="Select a doctor and define the reason of your visit" number="3" />
+                <StyledInputWrapper>
+                  <StyledLabel>Occupation</StyledLabel>
+                  <Field name="occupation" id="occupation" component={SelectOccupations} handleReset={setFieldValue} />
+                </StyledInputWrapper>
+                <StyledInputWrapper>
+                  <StyledLabel>Doctor’s Name</StyledLabel>
+                  <Field name="doctorName" id="doctorName" component={SelectDoctor} />
+                </StyledInputWrapper>
+                <StyledInputWrapper>
+                  <StyledLabel>Reason for the visit</StyledLabel>
+                  <Field name="reason" id="reason" type="text" component={Input} placeholder="Leave a reason" />
+                </StyledInputWrapper>
+                <StyledInputWrapper>
+                  <StyledLabel>Note</StyledLabel>
+                  <Field name="note" id="note" type="text" component={Input} placeholder="Leave a note if needed" />
+                </StyledInputWrapper>
+              </StyledContainer>
+            </StyledFormContainer>
+            <StyledBtnContainer>
+              <Button />
+            </StyledBtnContainer>
+          </StyledForm>
+        </StyledWrapper>
+      )}
+    </Formik>
   );
 };
