+++ src/store/user/patientOperations.ts
@@ -5,26 +5,7 @@ import { DoctorsBySpecializationType } from '../../components/models/doctorsBySp
 import { AddAppointmentResponseType, AddAppointmentType } from '../../components/models/addAppointment.model';
 import { NotifyError, NotifySuccess } from '../../components/UI/Notify';
 import { api } from '../../services/api';
-
-export const getAllSpecializations = () => api.get('specializations').then((response) => response.data.map((item: SpecializationsType) => (
-  {
-    value: item.id,
-    label: item.specialization_name,
-  }
-)));
-
-export const getDoctorsBySpecializations = (specializationId: string) => api.get(`doctors/specialization/${specializationId}`)
-  .then((response) => response.data.map((item: DoctorsBySpecializationType) => ({
-    value: item.id,
-    label: `${item.first_name} ${item.last_name}`,
-  })));
-
-export const getAvailableTime = (doctorId: string, date: string) => api.get('appointments/time/free', {
-  params: {
-    doctorId,
-    date,
-  },
-});
+import { ITimeSlots } from '../../components/models/TimeSlotsType.model';
 
 export const getSpecializations = () => async (dispatch: Dispatch<{ type: string }>) => {
   dispatch(patientActions.getSpecializationsRequest());
@@ -56,6 +37,23 @@ export const getDoctors = (specializationId: string) => async (
     dispatch(patientActions.getDoctorsSuccess(data));
   } catch (error) {
     dispatch(patientActions.getDoctorsError((error as Error).message));
+  }
+};
+
+export const getTimeSlots = (values: ITimeSlots) => async (
+  dispatch: Dispatch<{ type: string }>,
+) => {
+  dispatch(patientActions.getTimeSlotsRequest());
+  try {
+    const { data } = await api.get('appointments/time/free', {
+      params: {
+        doctorId: values.doctorId,
+        date: values.date,
+      },
+    });
+    dispatch(patientActions.getTimeSlotsSuccess(data));
+  } catch (error) {
+    dispatch(patientActions.getTimeSlotsError((error as Error).message));
   }
 };
